add get paths on staging gruntfile
